refactor: extract schema into its own file
