+++ next/src/components/codesLists/form/CodesListForm.tsx
@@ -1,6 +1,5 @@
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useBlocker, useNavigate } from '@tanstack/react-router';
-import i18next from 'i18next';
 import {
   type Control,
   Controller,
@@ -12,7 +11,6 @@ import {
   useForm,
 } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
-import { z } from 'zod';
 
 import DirtyStateDialog from '@/components/layout/DirtyStateDialog';
 import Button, { ButtonStyle } from '@/components/ui/Button';
@@ -28,6 +26,8 @@ import { type CodesList } from '@/models/codesLists';
 import { FormulasLanguages } from '@/models/questionnaires';
 import { Variable } from '@/models/variables';
 
+import { type FormValues, schema } from './schema';
+
 interface CodesListFormProps {
   /** In an update case, initial codes list value. */
   codesList?: Omit<CodesList, 'id'>;
@@ -39,34 +39,6 @@ interface CodesListFormProps {
   /** Function that will be called with form data when the user submit the form. */
   onSubmit: SubmitHandler<FormValues>;
 }
-
-const codesSchema = z.object({
-  value: z
-    .string()
-    .min(1, { error: i18next.t('codesList.form.mustProvideCodeValue') }),
-  label: z
-    .string()
-    .min(1, { error: i18next.t('codesList.form.mustProvideCodeLabel') }),
-  get codes() {
-    return z.array(codesSchema).optional();
-  },
-});
-
-const schema = z
-  .object({
-    label: z
-      .string()
-      .min(1, { error: i18next.t('codesList.form.mustProvideLabel') }),
-    codes: codesSchema
-      .array()
-      .min(1, { error: i18next.t('codesList.form.mustProvideCodes') }),
-  })
-  .check((ctx) => {
-    // Handle duplicate on code values
-    validateDuplicateValues(ctx.value.codes, 'codes', ctx);
-  });
-
-export type FormValues = z.infer<typeof schema>;
 
 /**
  * Create or edit a codes list.
@@ -348,60 +320,4 @@ function CodesField({
       ))}
     </>
   );
-}
-
-// Helper function to check for duplicate values and add errors
-function validateDuplicateValues(
-  codes: z.infer<typeof codesSchema>[],
-  pathPrefix: string,
-  ctx: z.core.ParsePayload<FormValues>,
-) {
-  const valuePaths: { value: string; paths: string[] }[] = [];
-
-  // Function to collect values and their paths recursively
-  const collectValues = (
-    codes: z.infer<typeof codesSchema>[],
-    pathPrefix: string,
-  ) => {
-    codes.forEach((code, index) => {
-      const currentPath = `${pathPrefix}.${index}.value`; // Path to the value field
-
-      // Add the value and path to the valuePaths array
-      const existingValue = valuePaths.find(
-        (item) => item.value === code.value,
-      );
-      if (existingValue) {
-        // If value exists, push the path
-        existingValue.paths.push(currentPath);
-      } else {
-        // If value does not exist, create a new entry
-        valuePaths.push({ value: code.value, paths: [currentPath] });
-      }
-
-      // If the code has subcodes, we add them recursively
-      if (code.codes?.length) {
-        collectValues(code.codes, `${pathPrefix}.${index}.codes`);
-      }
-    });
-  };
-
-  // Collect all values and their paths
-  collectValues(codes, pathPrefix);
-
-  // Check for duplicates and add validation issues for duplicate values
-  valuePaths.forEach(({ value, paths }) => {
-    if (paths.length > 1) {
-      // If the value appears more than once, it's a duplicate
-      paths.forEach((path) => {
-        ctx.issues.push({
-          code: 'custom',
-          message: i18next.t('codesList.form.mustProvideUniqueValue', {
-            value,
-          }),
-          input: ctx.value,
-          path: path.split('.'), // Add issue to all paths where the value appears
-        });
-      });
-    }
-  });
 }
