+++ next/src/components/ui/form/Field.tsx
@@ -1,3 +1,5 @@
+import React from 'react';
+
 import { Field as BaseUIField } from '@base-ui-components/react/field';
 import { FieldError } from 'react-hook-form';
 
@@ -12,7 +14,7 @@ export type Props = {
    * An accessible label that is automatically associated with the field
    * control.
    */
-  label: string;
+  label: string | React.ReactNode;
   /**
    * Identifies the field when a form is submitted. Takes precedence over the
    * `name` prop on the `<Field.Control>` component.
